damien-cli/pyproject.toml
[tool.poetry]
name = "damien-cli-project"
version = "0.1.0"
description = "A CLI tool to manage Gmail with LLM capabilities."
authors = ["ivan-rivera-projects <ivan.rivera.email@gmail.com>"]
license = "MIT"
readme = "README.md"

# Tells Poetry that the actual Python package code is in the 'damien_cli' directory.
packages = [
    { include = "damien_cli" }
]

[tool.poetry.dependencies]
# Constrain Python to compatible versions (avoid 3.13+ conflicts)
python = ">=3.11,<3.13"

# Main application dependencies are listed here.
click = ">=8.2.0,<9.0.0"
google-api-python-client = ">=2.169.0,<3.0.0"
google-auth-oauthlib = ">=1.2.2,<2.0.0"
google-auth-httplib2 = ">=0.2.0,<0.3.0"
pydantic = "^2.11.4"
pydantic-settings = "^2.3.4" # Added for BaseSettings

# AI Intelligence Layer Dependencies
openai = "^1.35.0"  # For GPT integration
tiktoken = "^0.7.0"  # For token counting
python-json-logger = "^2.0.7" # Added for structured logging

# ML and Data Science Dependencies (Core)
torch = "2.1.0"  # PyTorch CPU version - compatible with sentence-transformers
sentence-transformers = "2.6.0"  # For sentence embeddings - compatible version  
numpy = "^1.24.0,<2.0.0"  # Compatible with sentence-transformers and torch
scikit-learn = "^1.4.0"  # For clustering and ML algorithms
pandas = "^2.2.0"  # For data analysis and manipulation
tqdm = "^4.66.0"  # For progress bars
psutil = "^5.9.0"  # For system monitoring (performance metrics)

# Optional NLP Dependencies (can be installed manually if needed)
# nltk = "^3.8.1"  # For text processing - OPTIONAL
# spacy = "^3.7.0"  # For advanced NLP - OPTIONAL (causes conflicts)

# Optional Visualization Dependencies (can be installed manually if needed)  
# matplotlib = "^3.8.0"  # For visualization - OPTIONAL
# seaborn = "^0.13.0"  # For advanced plots - OPTIONAL

# Vector Storage (simplified version)
# chromadb = "^0.4.0"  # For vector storage - OPTIONAL (can add back if needed)
python-dotenv = "^1.1.0"
tenacity = "^9.1.2"

[tool.poetry.group.dev.dependencies]
# Development-specific dependencies.
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
black = "^25.1.0"
flake8 = "^7.2.0"
pytest-asyncio = "^0.23.0" # Added for running async tests

[tool.poetry.scripts]
damien = "damien_cli.cli_entry:damien"

[build-system]
# Standard build system configuration for Poetry projects.
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"
